Move content classes to new namespace
* creates a new `content` namespace and moves content-related classes
  there
* update tests and `views.py`
