+++ truetone/test/test_content_metadatum.py
@@ -1,10 +1,10 @@
 import os
-from truetone.models import content_metadatum
+from truetone.models.content.metadatum import ContentMetadatum
 
 
 def test_contains_a_slug_matching_the_file_name():
     cwd = os.path.dirname(os.path.realpath(__file__))
-    metadata = content_metadatum.ContentMetadatum.build(
+    metadatum_instance = ContentMetadatum.build(
             f'{cwd}/fixtures/content/path-1.md')
-    assert metadata.slug == 'path-1'
-    assert metadata.title == 'Heading 1'
+    assert metadatum_instance.slug == 'path-1'
+    assert metadatum_instance.title == 'Heading 1'
